+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [3.5.0] - 2023-05-16
 ### Added
-- Added support to pass in a git token on init
+- Added support to pass in a github token, username and password as named parameters
 
 ## [3.4.0] - 2023-02-24
 ### Added